Avoid package-level DSA backend imports
